+++ packages/list-box/src/vaadin-list-box.d.ts
@@ -5,8 +5,8 @@
  */
 import { ControllerMixin } from '@vaadin/component-base/src/controller-mixin.js';
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
-import { MultiSelectListMixin } from '@vaadin/vaadin-list-mixin/vaadin-multi-select-list-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { MultiSelectListMixin } from './vaadin-multi-select-list-mixin.js';
 
 /**
  * Fired when the `items` property changes.
