+++ packages/list-box/src/vaadin-multi-select-list-mixin.d.ts
@@ -4,7 +4,7 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import type { Constructor } from '@open-wc/dedupe-mixin';
-import type { ListMixinClass } from './vaadin-list-mixin.js';
+import type { ListMixinClass } from '@vaadin/component-base/src/list-mixin.js';
 
 /**
  * A mixin for `nav` elements, facilitating multiple selection of childNodes.
