+++ packages/tabs/test/typings/tabs.types.ts
@@ -1,4 +1,5 @@
 import '../../vaadin-tabs.js';
+import type { ListMixinClass } from '@vaadin/component-base/src/list-mixin.js';
 import type { ResizeMixinClass } from '@vaadin/component-base/src/resize-mixin.js';
 import type { TabsItemsChangedEvent, TabsSelectedChangedEvent } from '../../vaadin-tabs.js';
 
@@ -6,6 +7,7 @@ const tabs = document.createElement('vaadin-tabs');
 
 const assertType = <TExpected>(actual: TExpected) => actual;
 
+assertType<ListMixinClass>(tabs);
 assertType<ResizeMixinClass>(tabs);
 
 tabs.addEventListener('items-changed', (event) => {
