--- packages/vaadin-list-mixin/README.md
@@ -1,14 +1,0 @@
-# vaadin-list-mixin
-
-`vaadin-list-mixin` is a mixin for `nav` elements, facilitating navigation and selection of childNodes.
-
-[![npm version](https://badgen.net/npm/v/@vaadin/vaadin-list-mixin)](https://www.npmjs.com/package/@vaadin/vaadin-list-mixin)
-[![Discord](https://img.shields.io/discord/732335336448852018?label=discord)](https://discord.gg/PHmkCKC)
-
-## Contributing
-
-Read the [contributing guide](https://vaadin.com/docs/latest/contributing/overview) to learn about our development process, how to propose bugfixes and improvements, and how to test your changes to Vaadin components.
-
-## License
-
-Apache License 2.0
