--- packages/vaadin-list-mixin/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@vaadin/vaadin-list-mixin",
-  "version": "24.0.0-alpha4",
-  "publishConfig": {
-    "access": "public"
-  },
-  "description": "vaadin-list-mixin",
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/vaadin/web-components.git",
-    "directory": "packages/vaadin-list-mixin"
-  },
-  "author": "Vaadin Ltd",
-  "homepage": "https://vaadin.com/elements",
-  "bugs": {
-    "url": "https://github.com/vaadin/web-components/issues"
-  },
-  "main": "vaadin-list-mixin.js",
-  "module": "vaadin-list-mixin.js",
-  "type": "module",
-  "files": [
-    "*.d.ts",
-    "vaadin-*.js"
-  ],
-  "keywords": [
-    "Vaadin",
-    "vaadin-list-mixin",
-    "web-components",
-    "web-component",
-    "polymer"
-  ],
-  "dependencies": {
-    "@open-wc/dedupe-mixin": "^1.3.0",
-    "@polymer/polymer": "^3.0.0",
-    "@vaadin/component-base": "24.0.0-alpha4"
-  },
-  "devDependencies": {
-    "@esm-bundle/chai": "^4.3.4",
-    "@vaadin/testing-helpers": "^0.3.2",
-    "sinon": "^13.0.2"
-  }
-}
